Move `Docker` trait into a better place
The `Docker` and `DockerAsync` trait are necessary for the `Container`
and `ContainerAsync` structs to work properly. As such, they should
be defined in there because those structs motivate their existence.
