+++ src/core.rs
@@ -1,18 +1,18 @@
-pub(crate) use self::docker::Docker;
+pub(crate) use container::Docker;
+pub(crate) use container_async::DockerAsync;
+
 pub use self::{
     container::Container,
     container_async::ContainerAsync,
-    docker::RunArgs,
     image::{Image, Port, WaitFor},
 };
+pub use run_args::RunArgs;
 
 mod container;
 mod container_async;
-mod docker;
 pub mod env;
 mod image;
+mod run_args;
 
 pub(crate) mod logs;
 pub(crate) mod ports;
-
-pub(crate) use container_async::DockerAsync;
